perf(decoder): use primitive integers
To ensure we don't use unnecessarily boxed integers, we use `ToIntFunction` in
all of the `FieldDefs` components. We also create a `LeanEnumArrayIntMap` to
effectively do the same in the `LengthOffsetCache`. Also added some bounds
checks to `vendorIdsFromRange()`. We've seen this give a further 1.25x speed
improvement in eager `TCString.decode()`.

Semi-unrelated, but I also found a test case in `TCStringV2EncoderTest` that
fails when run between 23:00 and 00:00 UTC. I fixed it in accordance with the
comment in `TCStringEncoder.TCStringEncoderV2` whereby the last update time
is used in lieu of the creation time.