perf(decoder): leaner bit set with less invariant checking
Introduce `UnsafeLeanBitSet`, a stripped down flavor of standard library's
`BitSet` that doesn't do as much invariant checking. It requires the caller,
primarily just the code in this module, to always be aware of how much space
they've allocated for its storage, and to never request any operation on an
out-of-bounds index. If unsure, simply calling `ensureCapacity()` with a number
1 higher than any index a caller is about to operate on will always be safe.

This change tweaks `TCStringV1`, `TCStringV2`, and `LeanEnumArrayIntMap` to
leverage this new data structure. Testing this on our production workload seems
to give an extra 10 to 15 percent performance boost.